Refactored the shared code into isrcshared.py. I thought I could get by just importing from isrcsubmit.py but it has a binary dependency on the discid shared library so this causes problems, specifically on 64-bit windows.
